Added script snippet type support, and subdirectory for these is the convention
