+++ src/Syringe.Core/Tests/Scripting/ScriptSnippetType.cs
@@ -1,0 +1,7 @@
+﻿namespace Syringe.Core.Tests.Scripting
+{
+	public enum ScriptSnippetType
+	{
+		BeforeExecute
+	}
+}
